requirements.txt
asgiref==3.7.2
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
channels==3.0.5
channels-redis==4.1.0
constantly==23.10.4
cryptography==45.0.6
daphne==3.0.2
Django==4.2.5
djangorestframework==3.16.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
legacy-cgi==2.6.4
msgpack==1.1.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
redis==5.0.0
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.4.4
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
zope.interface==7.2
